CMS-1935: Rename and move helper functions
